feat(DEVING-50): 내가 만든 모임 리스트 get api 연결
